feat: add new endpoints

+++ routes/web.php
@@ -2,25 +2,49 @@
 
 // use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ListingController;
 
 // All listings
 Route::get('/', [ListingController::class, 'index']);
 
 // Create listings form
-Route::get("/listings/create", [ListingController::class, 'create']);
+Route::get("/listings/create", [ListingController::class, 'create'])->middleware('auth');
 
 // Store listings
-Route::post("/listings", [ListingController::class, 'store'])->name("create_job");
+Route::post("/listings", [ListingController::class, 'store'])->name("create_job")->middleware('auth');;
+
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Single listing
-Route::get("/listings/{listing}", [ListingController::class, 'show']);
+Route::get("/listings/{listing}", [ListingController::class, 'show'])->middleware('auth');;
 
 // Edit Submit to Update
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');;
 
 // Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');;
 
 // Edit listings form
-Route::get("/listings/{listing}/edit", [ListingController::class, 'edit']);
+Route::get("/listings/{listing}/edit", [ListingController::class, 'edit'])->middleware('auth');;
+
+
+/**
+ * User Session
+ */
+
+// Show Register/Create Form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+// Create new user
+Route::post('/users', [UserController::class, 'store']);
+
+// Logout User
+Route::post('/logout', [UserController::class, 'logout']);
+
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Log In User
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
